refactor(springboot-app): 파일 경로 수정
