+++ springboot-app/src/main/java/com/inha/springbootapp/global/config/GlobalCorsConfig.java
@@ -1,6 +1,5 @@
-package com.inha.springbootapp;
+package com.inha.springbootapp.global.config;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
